refactor of careerDetails
