--- src/components/CareersDetail/CareersData/index.tsx
@@ -1,1 +1,0 @@
-export { CareersDataContainer as CareersData } from "./container";
