+++ src/components/CareersSection/CareersDetail/index.tsx
