+++ src/components/CareersSection/CareersDetail/interfaces.ts
