+++ src/components/CareersSection/GraduateCareerDetail/index.tsx
@@ -1,12 +1,13 @@
 import React, { FunctionComponent } from "react";
 import { IApplicantCareer } from "$interfaces/Applicant";
-import { IGraduateTranslations } from "../CareersData/interfaces";
+import { IGraduateTranslations } from "../CareersDetail/interfaces";
 
 export const GraduateCareerDetail: FunctionComponent<IGraduateCareerDetail> = (
   {
     applicantCareer,
     translations
-  }) => (
+  }
+) => (
   <>
     <b>{translations.isGraduate} </b>
     {translations.connector}
