+++ src/components/CareersSection/component.tsx
@@ -2,19 +2,20 @@ import React, { FunctionComponent } from "react";
 import { ICareersProps } from "./interface";
 import styles from "./styles.module.scss";
 import { Subtitle } from "$components/Subtitle";
-import { CareersData } from "./CareersData";
+import { CareersDetail } from "./CareersDetail";
 
-export const CareersDetail: FunctionComponent<ICareersProps> = (
+export const CareersSection: FunctionComponent<ICareersProps> = (
   {
     className,
     careers,
     translations
-  }) => {
+  }
+) => {
   return (
     <div className={className}>
       <section className={styles.items}>
         <Subtitle>{translations.careersTitle}</Subtitle>
-        <CareersData careers={careers} className={styles.item}/>
+        <CareersDetail careers={careers} className={styles.item}/>
       </section>
     </div>
   );
