+++ src/components/CareersSection/index.tsx
@@ -1,0 +1,1 @@
+export { CareersSectionContainer as CareersSection } from "./container";
